changed error conditions
