+++ subject
@@ -1,0 +1,27 @@
+4
+##start
+1 5 6
+##end
+8 7 7
+2 4 6
+3 8 5
+4 1 2
+5 9 8
+6 9 9
+7 0 7
+#    1
+#  / |
+# 2  4--5
+# |  |  |
+# 3--6  7
+#    | /
+#    8
+1-2
+1-4
+4-5
+4-6
+5-7
+2-3
+3-6
+7-8
+6-8
